Changed autoId from int to long (missed one spot). Have now completed this change.

+++ app/src/main/java/albright/csit/legosetstracker/LegoSet.java
@@ -26,7 +26,7 @@ public class LegoSet {
 
   ////  Setters
   /////////////////////////////////////
-  public void setAutoId(int autoId) {
+  public void setAutoId(long autoId) {
     this.autoId = autoId;
   }
